requirements.txt
langchain
faiss-cpu
sentence-transformers
openai
unstructured
streamlit
